sam/User-wallet-sam/streams-dynamodb/requirements.txt
aws-kinesis-agg
amazon.ion
boto3